refactor: setup pkg as its own module
