+++ pkg/agave/assets/install.sh
