+++ pkg/runner/assets/lib.bash
