+++ pkg/solana/assets/genesis.sh
