+++ pkg/ssh/connection.go
