+++ pkg/ssh/ready.go
