+++ provider/pkg/svm/genesis.go
@@ -1,7 +1,7 @@
 package svm
 
 import (
-	"github.com/abklabs/pulumi-svm/provider/pkg/ssh"
+	"github.com/abklabs/svmkit/pkg/ssh"
 )
 
 // GenesisArgs represents the arguments required to configure a genesis ledger.
