+++ sdk/dotnet/version.txt
@@ -1,1 +1,1 @@
-0.0.1-alpha.1726480324+c75f60e1.dirty
+0.0.1-alpha.1726494044+02eb4bef.dirty
